Permission

+++ app/Http/Controllers/EventController.php
@@ -2,36 +2,37 @@
 
 namespace App\Http\Controllers;
 
+use Illuminate\Support\Facades\Auth;
 use App\Event;
 use App\Organize;
 use Illuminate\Http\Request;
 
 class EventController extends Controller
 {
-    public function listEvent(){
-
-
-        $this->authorize("viewAny");
+    public function index(){
+        $this->authorize("view" , Event::class);
+        $organizations=Organize::all();
         $events =Event::paginate(20);
         return view("admin.event.list", [
-            "events" => $events
-//
-//
+            "events" => $events,
+            "organizations"=>$organizations,
         ]);
-
     }
 
-    public function newEvent()
+    public function new()
     {
+        $this->authorize("new",Event::class);
+        $events=Event::paginate(20);
         $organizations=Organize::all();
         return view("admin.event.new",
         [
+            "events" => $events,
             "organizations"=>$organizations,
-
         ]);
     }
 
-    public function saveEvent(Request $request){
+    public function save(Request $request){
+        $this->authorize("save",Event::class);
         $request->validate([
             "title"=>"required",
             "image"=>"required",
@@ -41,7 +42,6 @@ class EventController extends Controller
             "end_at"=>"required",
             "total_money"=>"required",
             "organization_id"=>"required"
-
         ]);
         try {
             Event::create([
@@ -52,17 +52,19 @@ class EventController extends Controller
                 "start_at"=>$request->get("start_at"),
                 "end_at"=>$request->get("end_at"),
                 "total_money"=>$request->get("total_money"),
-                "organization_id"=>$request->get("organization_id")
+                "organization_id"=>$request->get("organization_id"),
+                "user_id" => Auth::id()
             ]);
 
         }catch (\Exception $exception){
             return $exception->getMessage();
         }
-       return redirect()->to("/admin/list-event");
+       return redirect()->to("/admin/events/list");
     }
 
-    public function editEvent($id){
+    public function edit($id){
         $event = Event::findOrFail($id);
+        $this->authorize("edit",$event, Event::class);
         $organizations=Organize::all();
         return view("admin.event.edit",[
             "event"=>$event,
@@ -70,8 +72,9 @@ class EventController extends Controller
             ]);
     }
 
-    public function updateEvent($id,Request $request){
+    public function update($id,Request $request){
         $event = Event::findOrFail($id);
+        $this->authorize("update",$event,Event::class);
         $request->validate([
             "title"=>"required",
             "image"=>"required",
@@ -96,16 +99,18 @@ class EventController extends Controller
         }catch (\Exception $exception){
             return redirect()->back();
         }
-        return redirect()->to("/admin/list-event");
+        return redirect()->to("/admin/events/list");
     }
 
-    public function deleteEvent($id){
+    public function delete($id){
         $event = Event::findOrFail($id);
+        $this->authorize("delete",$event,Event::class);
+
         try {
             $event->delete();
         }catch (\Exception $exception){
 
         }
-        return redirect()->to("admin/list-event");
+        return redirect()->to("admin/events/list");
     }
 }
